requirements.txt
python==3.6.13
mysql-connector-python==8.0.26
nltk==3.6.2
requests==2.26.0
torch==1.9.0
torchtext==0.10.0